Split test suite into unit/ and integration/ subdirectories
- Move test_imports, test_config, test_cli, test_time_slicing,
  test_train_model_warnings into tests/unit/
- Refactor test_clamping to use DummyDatastore instead of
  MDPDatastore and move to tests/unit/
- Move test_datastores, test_graph_creation, test_training,
  test_datasets, test_plotting, test_plot_graph into tests/integration/
- Add tests/integration/conftest.py with S3 fixture and
  init_datastore_example; strip tests/conftest.py to shared env setup
- Add tests/unit/conftest.py with no S3 dependencies
- Register pytest marker 'integration' and testpaths in pyproject.toml
- Split CI pytest step into unit and integration runs
