+++ tests/integration/test_graph_creation.py
@@ -10,7 +10,7 @@ import torch
 from neural_lam.create_graph import create_graph_from_datastore
 from neural_lam.datastore import DATASTORES
 from neural_lam.datastore.base import BaseRegularGridDatastore
-from tests.conftest import init_datastore_example
+from tests.integration.conftest import init_datastore_example
 
 
 @pytest.mark.parametrize("graph_name", ["1level", "multiscale", "hierarchical"])
